+++ docs/changelog.md
@@ -13,6 +13,7 @@ Click on the version in the following list to see all changes since the previous
 
 * docs: move changelog into online documentation
 * docs: improve changelog with additional description
+* fix: do not rely on commons-lang 3.5 for checking numeric values
 
 ### [3.11.0] - 2019-03-30